handling updates from notes and properly displaying them
therfore added an OnUpdateCallback interface. It's update() method will
be called by the note, if one of the parameters of the note have been
changed.

the displays will be listening for this and update their visuals
accordingly.

also added a flag for the NoteDisplayView to save the IsInitializedState
to not have to check for null-objects but rather isInitialized() which
is a better check - maybe should add a statemachine for this with a
proper state