extended filtering and implemented sorting of notes
notes can now be sorted. for that a sortableNote has been introduced.
it contains a map of category and action. That action shall implement a
getData method, which will return the data corresponding with the given
category.

upon sortrequest of that specific category, getData of the correct
handler will be called and therfore return that specific value.

furthermore a test for the first sort mechanism sortByTitle has been
introduced. TO be able to run it, the logger needed to be modified.

The logger needs to be seperated from the android specific stuff, so
unit tests don't need to be run on an acutal machine, but on the
developers...