fixed issue with swipe in edit note fragment
fixed issue where swiping in edit task note fragment was not working
properly. issue was that the data was loaded twice and therfore adding
the views twice. this lead to an issue where one view was invisible, and
therefore initialized the values incorrectly.
calling loadvalue now only once and making sure to replace fragments
instead of just adding them

AB#29